Improve ModSelector; some cleanup
- show value in slider text
- hide slider if value is fixed
- show text with value above sliders for stats with multiple values
  (e.g. "Adds # to # ... Damage") and for stats with fixed or no values
- cleanup ModSelector, OverlaySlider and a bit of CraftWindow

+++ WPFSKillTree/Utils/Extensions/NumberExtensions.cs
@@ -4,7 +4,7 @@ namespace POESKillTree.Utils.Extensions
 {
     public static class NumberExtensions
     {
-        public static bool AlmostEquals(this float x, float y, float delta)
+        public static bool AlmostEquals(this float x, float y, double delta)
         {
             return Math.Abs(x - y) < delta;
         }
